Probe for more up to date noload RTT

+++ concurrency-limits-core/src/main/java/com/netflix/concurrency/limits/Limit.java
@@ -17,12 +17,12 @@ public interface Limit {
         /**
          * @return Maximum number of inflight observed during the sample window
          */
-        long getMaxInFlight();
+        int getMaxInFlight();
         
         /**
          * @return Number of observed RTTs in the sample window
          */
-        long getSampleCount();
+        int getSampleCount();
         
         /**
          * @return True if there was a timeout
